Add script to restore modification date of files
When git checks out files, it sets the modification time to the current
time. Since the tool to create the index file uses the modification time
of the file, we need the date when the file was committed to the repository,
not the current date. This script restores the modification time of all
files under version control to the date of their commit.